requirements.txt
Flask==1.1.2
h5py==2.10.0
numpy==1.19.4
gunicorn==20.0.4
Pillow==8.0.1
tensorflow-cpu==2.3.0
scikit-image==0.18.0 